comet层添加rpc方法

+++ libs/proto/router.go
@@ -1,7 +1,7 @@
 package proto
 
 type Router struct {
-	ServerId int16
+	ServerId int8
 	RoomId   int32
 	UserId   int64
 }
